backend-flask/requirements.txt
flask
flask-cors
python-dotenv
requests
groq